files-api/pyproject.toml
[project]
name = "files-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "avr2002", email = "avr13405@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "aws-embedded-metrics>=3.3.0",
    "aws-xray-sdk>=2.14.0",
    "boto3>=1.38.32",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "openai>=1.84.0",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
api = ["moto[server]>=5.1.5"]
aws-lambda = ["mangum>=0.19.0"]
notebooks = ["ipykernel>=6.29.5", "jupyter>=1.1.1", "rich>=14.0.0"]
static-code-qa = ["pre-commit>=4.2.0", "ruff>=0.11.13"]
stubs = ["boto3-stubs[cloudwatch,logs,s3]>=1.38.32"]
test = [
    "locust>=2.37.9",
    "moto[s3]>=5.1.5",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "uvicorn>=0.34.3",
]
docker = [
    "uvicorn>=0.34.3",
]
dev = ["files-api[aws-lambda,test,release,static-code-qa,stubs,notebooks,api]"]

[tool.poe]
executor.type = "uv"

[tool.poe.tasks.clean]
help = "Remove all files generated by tests, builds, or operating this codebase"
shell = """
rm -rf dist build coverage.xml test-reports sample/ tests/artifacts/

find . \\
    -type d \\
    \\( \\
    -name "*cache*" \\
    -o -name "tests/artifacts" \\
    -o -name "*.dist-info" \\
    -o -name "*.egg-info" \\
    -o -name "*htmlcov" \\
    -o -name "*.metaflow" \\
    -o -name "*.metaflow.s3" \\
    -o -name "*.mypy_cache" \\
    -o -name "*.pytest_cache" \\
    -o -name "*.ruff_cache" \\
    -o -name "*__pycache__" \\
    \\) \\
    -not -path "*env/*" \\
    -exec rm -r {} + || true

find . \\
    -type f \\
    -name "*.pyc" \\
    -o -name "*.DS_Store" \\
    -o -name "*.coverage*" \\
    -not -path "*env/*" \\
    -exec rm {} +
"""

[tool.poe.tasks.build]
help = "Build a wheel and sdist from the Python source code"
shell = "bash run.sh build"

[tool.poe.tasks.update-dashboard]
help = "Update the CloudWatch dashboard with new version information"
shell = "bash run.sh update-dashboard"

[tool.poe.tasks.deploy-lambda]
help = "Package and deploy the src/ code by updating an existing AWS Lambda function"
shell = "bash run.sh deploy-lambda"
deps = ["clean"]

[tool.poe.tasks.deploy-lambda-code]
help = "Deploy only the application code to the lambda function"
shell = "bash run.sh deploy-lambda:code"
deps = ["clean"]

[tool.poe.tasks.install-generated-sdk]
help = "Install the generated SDK in newly created venv"
shell = "bash run.sh install-generated-sdk"

[tool.poe.tasks.generate-client-library]
help = "Generate Python client library from OpenAPI spec"
shell = "bash run.sh generate-client-library"

[tool.poe.tasks.run]
help = "Start the FastAPI app locally with real AWS credentials"
shell = "bash run.sh run"

[tool.poe.tasks.run-docker]
help = "Start the FastAPI app in a Docker container"
shell = "bash run.sh run-docker"

[tool.poe.tasks.run-locust]
help = "Run load tests with Locust"
shell = "bash run.sh run-locust"

[tool.poe.tasks.run-mock]
help = "Start the FastAPI app with mocked AWS and OpenAI endpoints"
shell = "bash run.sh run-mock"

[tool.poe.tasks.run-local]
help = "Start the FastAPI app locally with AWS & OpenAI credentials"
shell = "bash run.sh run-local"

[tool.poe.tasks.help]
help = "Print available tasks"
shell = "bash run.sh help"

[tool.poe.tasks.install]
help = "Install core and development Python dependencies"
shell = "bash run.sh install"

[tool.poe.tasks.lint]
help = "Run linting, formatting, and other static code quality tools"
shell = "bash run.sh lint"

[tool.poe.tasks.lint-ci]
help = "Run lint checks in CI environment"
shell = "bash run.sh lint:ci"

[tool.poe.tasks.publish-prod]
help = "Publish package to PyPI production"
shell = "bash run.sh publish:prod"

[tool.poe.tasks.publish-test]
help = "Publish package to TestPyPI"
shell = "bash run.sh publish:test"

[tool.poe.tasks.release-prod]
help = "Release to production (includes linting, testing, and publishing)"
shell = "bash run.sh release:prod"

[tool.poe.tasks.release-test]
help = "Release to test environment (includes linting, testing, and publishing)"
shell = "bash run.sh release:test"

[tool.poe.tasks.serve-coverage-report]
help = "Serve the HTML test coverage report on localhost:8000"
shell = "bash run.sh serve-coverage-report"

[tool.poe.tasks.test-ci]
help = "Execute tests against the installed package in CI environment"
shell = "bash run.sh test:ci"

[tool.poe.tasks.test-quick]
help = "Execute tests that are not marked as 'slow'"
shell = "bash run.sh test:quick"

[tool.poe.tasks.test]
help = "Execute all tests"
shell = "bash run.sh run-tests"

[tool.poe.tasks.test-wheel-locally]
help = "Test the built wheel package locally in a fresh environment"
shell = "bash run.sh test:wheel-locally"
